JDBC.java phoneNum parsed as string
I still haven't managed to get JDBC.java to function.  However, there was a MismatchException being thrown that I was able to resolve.  The program doesn't throw any errors on my end.  However, it appears to be stuck in an infinite while loop and never inputs the data into the DB.